Feito o o metodo de adição e remoção total no service e o metodo de adição na controller, metodos de conversão na DTO e ajuste de tipo de atributo bolsa

+++ portal/VagaEstagio/VagaEstagio/src/main/java/VagaEstagio/model/VagaModel.java
@@ -7,6 +7,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "tb__vaga")
 @Getter
@@ -22,7 +24,7 @@ public class VagaModel {
     private String descricao;
 
     @Column(nullable = false)
-    private Integer bolsa;
+    private BigDecimal bolsa;
 
 
     @Column(nullable = false) @Enumerated(EnumType.STRING)
